refactor: rename use-mobile to useMobile
