+++ src/components/shadcn/sidebar.tsx
@@ -3,7 +3,7 @@ import { Slot } from '@radix-ui/react-slot'
 import { VariantProps, cva } from 'class-variance-authority'
 import { PanelLeftIcon } from 'lucide-react'
 
-import { useIsMobile } from '@/hooks/use-mobile'
+import { useIsMobile } from '@/hooks/useMobile'
 import { cn } from '@/lib/utils'
 import { Button } from '@/components/shadcn/button'
 import { Input } from '@/components/shadcn/input'
