+++ src/components/shared/Sidebar.tsx
@@ -7,7 +7,7 @@ import Notification from '../../assets/icons/Notification'
 import Settings from '../../assets/icons/Settings'
 import SidebarToggle from '../../assets/icons/SidebarToggle'
 import Tracker from '@/assets/icons/Tracker'
-import { useIsMobile } from '@/hooks/use-mobile'
+import { useIsMobile } from '@/hooks/useMobile'
 import Logo from '../../assets/icons/Logo'
 import userProfile from '../../assets/images/ix-user-profile-filled.svg'
 import { UserRole } from '../../util/interfaces'
